I made the App file simpler by taking the code to another file to work in
